requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.18.4
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via starlette
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via
    #   rq
    #   uvicorn
coverage[toml]==7.13.4
    # via pytest-cov
croniter==6.0.0
    # via rq
fastapi==0.135.1
    # via -r requirements.in
greenlet==3.3.2
    # via
    #   -r requirements.in
    #   sqlalchemy
gunicorn==25.1.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
iniconfig==2.3.0
    # via pytest
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
packaging==26.0
    # via
    #   gunicorn
    #   pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==7.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via croniter
python-dotenv==1.2.2
    # via -r requirements.in
pytz==2026.1.post1
    # via croniter
redis==7.3.0
    # via
    #   -r requirements.in
    #   rq
requests==2.32.5
    # via -r requirements.in
rq==2.7.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.48
    # via
    #   -r requirements.in
    #   alembic
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via requests
uvicorn==0.41.0
    # via -r requirements.in
